added debug menu for level load and effects

+++ entitites/ball/ball.cs
@@ -4,7 +4,7 @@ using System;
 public partial class ball : CharacterBody2D
 {
 	private int speed = 500;
-	private float piercing = 0;
+	private int piercing = 0;
 	private Vector2 velocity;
 	private Vector2 startPosition;
 	private Vector2 previousPosition;
@@ -19,7 +19,7 @@ public partial class ball : CharacterBody2D
 		var collisionInfo = MoveAndCollide(velocity * (float)delta);
         if (collisionInfo != null){
 			if (collisionInfo.GetCollider().GetType()==typeof(paddle)){
-				piercing = (float)collisionInfo.GetCollider().GetIndexed("Piercing");
+				piercing = (int)collisionInfo.GetCollider().GetIndexed("Piercing");
 				//forms a vector pointing from the paddle to the ball. The further out from the center of the paddle, the wider the ball will go.
 				Vector2 relativeVector = (GlobalPosition - (Vector2)collisionInfo.GetCollider().GetIndexed("global_position")).Normalized()*speed;
 				//mix of standard bounce collision and the relative vector. Adjust the mix to change bounce behaviour.
